feat(): Listeners for state changes and progress updates of devices are now called in a non-realtime thread.
We intend to execute the device controllers in a realtime thread, so
that nothing can disturb the printing process. But we don't want to call
the listeners of state changes and progress update on a realtime thread.
Therefore we had to separate the state setting from the listener
calling. See Device.hh.